@@ -11,16 +11,16 @@ authors = ["Keith Gaughan <keith@gaughan.ie>"]
 license = "MIT"
 
 [tool.poetry.dependencies]
-python = "^3.7"
+python = "^3.8"
 
 [tool.poetry.dev-dependencies]
-black = "^20.8b1"
+black = "^22.3.0"
 coverage = "^5.5"
-flake8 = "^3.9.0"
-flake8-bugbear = "^21.3.2"
-isort = "^5.8.0"
+flake8 = "^3.9.2"
+flake8-bugbear = "^21.11.29"
+isort = "^5.10.1"
 pep8-naming = "^0.11.1"
-pylint = "^2.7.4"
+pylint = "^2.15.5"
 pytest = "^6.2.4"
 pytest-cov = "^2.12.1"
 
